# Conflicts:
#	src/attendance/automation/AttendanceAutomation.java
#	src/attendance/automation/be/Class.java
#	src/attendance/automation/bll/AAManager.java
#	src/attendance/automation/gui/controller/CalendarViewController.java
#	src/attendance/automation/gui/controller/LoginViewController.java
#	src/attendance/automation/gui/controller/TeacherMainViewController.java

+++ src/attendance/automation/gui/controller/StudentMainViewController.java
@@ -96,7 +96,7 @@ public class StudentMainViewController implements Initializable {
 
   private void calculateAttendanceRate() throws DALException {
     //String string = "Total attendance " + (int) (manager.attendanceRate(st) * 100) + "%";
-    String string = "Total attendance " + (int) (aamodel.attendanceRate(st) * 100) + "%";
+    String string = "Total attendance " + (aamodel.attendanceRate(st) * 100) + "%";
     attendanceRate.setText(string);
   }
